+++ pkg/goctx/testdata/golden/TestE2E_Handle_TestFiles_CallsitesInTestsForProdFunc/util_test.go
@@ -1,0 +1,8 @@
+package main
+
+import "testing"
+
+func TestProdFunc(t *testing.T) {
+	ctx := t.Context()
+	_ = ProdFunc(ctx)
+}