+++ pkg/goctx/testdata/input/TestE2E_Handle_TestFiles_CallsitesInTestsForProdFunc/util.go
@@ -1,0 +1,6 @@
+package main
+
+// ProdFunc is a production function used by tests.
+func ProdFunc() int {
+	return 7
+}