+++ pkg/goctx/testdata/input/TestE2E_Handle_TestFiles_CallsitesInTestsForProdFunc/util_test.go
@@ -1,0 +1,7 @@
+package main
+
+import "testing"
+
+func TestProdFunc(t *testing.T) {
+	_ = ProdFunc()
+}