+++ pkg/goctx/testdata/input/TestE2E_Handle_TestFiles_TargetIsTestAndCallsitesInTests/helper_test.go
@@ -1,0 +1,6 @@
+package main
+
+// HelperTarget is defined in a _test.go file and is called by other test code.
+func HelperTarget() int {
+	return 42
+}