+++ pkg/goctx/testdata/input/TestE2E_Handle_TestFiles_TargetIsTestAndCallsitesInTests/util_test.go
@@ -1,0 +1,7 @@
+package main
+
+import "testing"
+
+func TestUsesHelper(t *testing.T) {
+	_ = HelperTarget()
+}