+++ pkg/goctx/testdata/input/TestE2E_Methods_Propagation/2/main.go
@@ -1,5 +1,7 @@
 package main
 
+import "context"
+
 type S struct{}
 
 func (s *S) target() {}