+++ pkg/goctx/testdata/input/TestE2E_ModuleWide_FromSubdir/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"example.com/e2e/sub"
+)
+
+func main() {
+	sub.FuncInNeedOfContext()
+}