+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2025-12-01
+
+### Added
+
+- Regression tests for the fixes contained in this release (see below).
+
+### Fixed
+
+- When you are in a subdir *below* where the `go.mod` is, and you run `goctx` with a TARGET argument that specifies a function in a source file in the current dir - but that function is called elsewhere in the module *outside* of this subdir - the app now adjusts all the relevant call sites within the module, not just the ones in the current subdir.
+
+- Both functions defined in `*_test.go` files, and callsites contained in `*_test.go` files (even in cases where no signatures of functions defined in `*_test.go` were changed), are now correctly handled.
+
+- App now correctly distinguishes between `MyFunc()`, `xyz.MyFunc()` (where `xyz` is a package name), `a.MyFunc()` (where `a` is an object of type `TypeA`), and `b.MyFunc()` (where `b` is an object of type `TypeB`). Changes to the function signature of one of these will not affect call sites where the others are called.
+
+- When a function `MyFunc` contains calls to two functions `MyOtherFunc1` and `MyOtherFunc2`, and a context argument has been added to both `MyOtherFunc1` and `MyOtherFunc2`, only one context argument will be added to `MyFunc`'s signature (and that single argument will be used in the calls to both `MyOtherFunc1` and `MyOtherFunc2`).
+
+### Changed
+
+- Structure and dir- and file-naming under `pkg/goctx/testdata/{input,golden}` has been organized & rationalized. In particular, subdirs of `pkg/goctx/testdata/input` and `pkg/goctx/testdata/golden` are now named identically to the test in which they are used.
+
 ## [0.11.4] - 2025-12-01
 
 ### Added
@@ -210,7 +230,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Printing of what next git tag _would_ be in Makefile `test` target, and associated flags in `actions/checkout@v4` step in `checks` workflow.
+- Printing of what next git tag *would* be in Makefile `test` target, and associated flags in `actions/checkout@v4` step in `checks` workflow.
 
 ## [0.5.2] - 2025-11-02
 
@@ -228,7 +248,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Print message about what next git tag _would_ be in Makefile `test` target.
+- Print message about what next git tag *would* be in Makefile `test` target.
 
 ### Fixed
 
@@ -272,7 +292,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.0] - 2025-10-30
 
-[unreleased]: https://github.com/preminger/goctx/compare/v0.11.4...HEAD
+[unreleased]: https://github.com/preminger/goctx/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/preminger/goctx/compare/v0.11.4...v0.12.0
 [0.11.4]: https://github.com/preminger/goctx/compare/v0.11.3...v0.11.4
 [0.11.3]: https://github.com/preminger/goctx/compare/v0.11.2...v0.11.3
 [0.11.2]: https://github.com/preminger/goctx/compare/v0.11.1...v0.11.2